// src/api/models/Tokens.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Jaipur
 * API pour le jeu de société Jaipur
 *
 * The version of the OpenAPI document: 1.0
 * Contact: julien@usson.me
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface Tokens
 */
export interface Tokens {
    /**
     * 
     * @type {Array<number>}
     * @memberof Tokens
     */
    diamonds: Array<number>;
    /**
     * 
     * @type {Array<number>}
     * @memberof Tokens
     */
    gold: Array<number>;
    /**
     * 
     * @type {Array<number>}
     * @memberof Tokens
     */
    silver: Array<number>;
    /**
     * 
     * @type {Array<number>}
     * @memberof Tokens
     */
    cloth: Array<number>;
    /**
     * 
     * @type {Array<number>}
     * @memberof Tokens
     */
    spice: Array<number>;
    /**
     * 
     * @type {Array<number>}
     * @memberof Tokens
     */
    leather: Array<number>;
}

export function TokensFromJSON(json: any): Tokens {
    return TokensFromJSONTyped(json, false);
}

export function TokensFromJSONTyped(json: any, ignoreDiscriminator: boolean): Tokens {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'diamonds': json['diamonds'],
        'gold': json['gold'],
        'silver': json['silver'],
        'cloth': json['cloth'],
        'spice': json['spice'],
        'leather': json['leather'],
    };
}

export function TokensToJSON(value?: Tokens | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'diamonds': value.diamonds,
        'gold': value.gold,
        'silver': value.silver,
        'cloth': value.cloth,
        'spice': value.spice,
        'leather': value.leather,
    };
}